docker/webapp_proxied/django_uwsgi/requirements.txt
# Core Libs
uwsgi == 2.0.12
django == 1.9.2

# Drivers 
psycopg2 == 2.6.1
mysqlclient == 1.3.7
cassandra-driver == 3.1.1

# Other Libs

